requirements.txt
fastapi
uvicorn[standard]
httpx
beautifulsoup4
trafilatura
cachetools
spacy
textstat
sentence-transformers
torch
google-generativeai
openai
pydantic>=2.0
python-dotenv
pytest
pytest-asyncio
scikit-learn
numpy
mypy
